// src/api/api.ts
export const EventType = {
    BAD: 'BAD',
    GOOD: 'GOOD',
    SUPERBAD: 'SUPERBAD',
} as const;

export type EventType = (typeof EventType)[keyof typeof EventType];
// ...
export interface GameEventDto {
    /**
     * Уникальный идентификатор события
     * @type {string}
     * @memberof GameEventDto
     */
    id: string;
    /**
     * Тип события
     * @type {EventType}
     * @memberof GameEventDto
     */
    type: EventType;
    /**
     * Описание события
     * @type {string}
     * @memberof GameEventDto
     */
    description: string;
    /**
     * Прибыли и влияние проекта (опционально)
     * @type {ProfitsDto}
     * @memberof GameEventDto
     */
    profits?: ProfitsDto;
}
// ...
export interface ProfitsDto {
    /**
     * Влияние на скорость разработки
     * @type {number}
     * @memberof ProfitsDto
     */
    speed: number;
    /**
     * Влияние на технический долг
     * @type {number}
     * @memberof ProfitsDto
     */
    techDebt: number;
    /**
     * Влияние на зарплату/бюджет
     * @type {number}
     * @memberof ProfitsDto
     */
    cash: number;
}
/**
 *
 * @export
 * @interface ProjectDto
 */
export interface ProjectDto {
    /**
     * Уникальный идентификатор проекта
     * @type {string}
     * @memberof ProjectDto
     */
    id: string;
    /**
     * Название проекта
     * @type {string}
     * @memberof ProjectDto
     */
    name: string;
    /**
     * Описание проекта
     * @type {string}
     * @memberof ProjectDto
     */
    description: string;
    /**
     * Тип проекта
     * @type {string}
     * @memberof ProjectDto
     */
    type: ProjectDtoTypeEnum;
    /**
     * Стоимость проекта
     * @type {number}
     * @memberof ProjectDto
     */
    cost: number;
    /**
     * Прибыли и влияние проекта (опционально)
     * @type {ProfitsDto}
     * @memberof ProjectDto
     */
    profits?: ProfitsDto;
    /**
     * Список возможных событий для данного проекта
     * @type {Array<GameEventDto>}
     * @memberof ProjectDto
     */
    events: Array<GameEventDto>;
}

export const ProjectDtoTypeEnum = {
    product: 'product',
    marketing: 'marketing',
    techDebt: 'techDebt',
} as const;

export type ProjectDtoTypeEnum = (typeof ProjectDtoTypeEnum)[keyof typeof ProjectDtoTypeEnum];